tegra: cpu: fix compiler warning arch/arm/mach-tegra/cpu-tegra3.c: In function 'min_cpus_notify': arch/arm/mach-tegra/cpu-tegra3.c:311:186: warning: comparison of distinct pointer types lacks a cast [enabled by default

+++ arch/arm/mach-tegra/cpu-tegra3.c
@@ -309,7 +309,7 @@ static int min_cpus_notify(struct notifier_block *nb, unsigned long n, void *p)
 	if ((n >= 1) && is_lp_cluster()) {
 		/* make sure cpu rate is within g-mode range before switching */
 		unsigned int speed = max(
-			tegra_getspeed(0), clk_get_min_rate(cpu_g_clk) / 1000);
+			tegra_getspeed(0), (unsigned int)clk_get_min_rate(cpu_g_clk) / 1000);
 		tegra_update_cpu_speed(speed);
 
 		if (!clk_set_parent(cpu_clk, cpu_g_clk)) {
